Bug fixing

+++ Code/GLCMFeatures2DFullMerge.h
@@ -16,11 +16,11 @@ class GLCMFeatures2DFullMerge : GLCMFeatures<T,R>  {
      private:
         GLCMFeatures<T, R> glcmComb;
         string normGLCM;
-        vector<double> actualSpacing;
+        vector<float> actualSpacing;
         void extractGLCMDataFullMerge(vector<T> &glcmData, GLCMFeatures2DFullMerge<T, R> glcmFeatures);
-        void fill2DMatrices(boost::multi_array<T, R> inputMatrix, boost::multi_array<double, 2> &glcMatrix, int depth, int angle);
+        void fill2DMatrices(boost::multi_array<T, R> inputMatrix, boost::multi_array<float, 2> &glcMatrix, int depth, int angle);
         std::vector<std::pair<T, T> > getNeighbours2D(boost::multi_array<T, R> inputMatrix, int depth, int directionX, int directionY);
-        boost::multi_array<double, 2> calculateMatrix2DFullMerge( boost::multi_array<T, R> inputMatrix, float maxIntensity);
+        boost::multi_array<float, 2> calculateMatrix2DFullMerge( boost::multi_array<T, R> inputMatrix, float maxIntensity);
 
         int N_g;
         vector<T> diagonalProbabilities;
@@ -37,7 +37,7 @@ class GLCMFeatures2DFullMerge : GLCMFeatures<T,R>  {
 		}
 		~GLCMFeatures2DFullMerge() {
 		}
-        void calculateAllGLCMFeatures2DFullMerge(GLCMFeatures2DFullMerge<T,R> &glcmFeat, boost::multi_array<T,R> inputMatrix, float maxIntensity, vector<double> spacing, ConfigFile config);
+        void calculateAllGLCMFeatures2DFullMerge(GLCMFeatures2DFullMerge<T,R> &glcmFeat, boost::multi_array<T,R> inputMatrix, float maxIntensity, vector<float> spacing, ConfigFile config);
         void writeCSVFileGLCM2DFullMerge(GLCMFeatures2DFullMerge<T,R> glcmFeat, string outputFolder);
 		void writeOneFileGLCM2DFullMerge(GLCMFeatures2DFullMerge<T, R> glcmFeat, string outputFolder);
 };
@@ -51,8 +51,8 @@ for the occurence of every neighbor pair.
 @param[out]: GLCM-matrix
 */
 template <class T, size_t R>
-boost::multi_array<double, 2> GLCMFeatures2DFullMerge<T,R>::calculateMatrix2DFullMerge( boost::multi_array<T, R> inputMatrix, float maxIntensity){
-    typedef boost::multi_array<double, 2> glcmat;
+boost::multi_array<float, 2> GLCMFeatures2DFullMerge<T,R>::calculateMatrix2DFullMerge( boost::multi_array<T, R> inputMatrix, float maxIntensity){
+    typedef boost::multi_array<float, 2> glcmat;
     int sizeMatrix = maxIntensity;
     int ang;
 	
@@ -95,7 +95,7 @@ As next step, the GLCMatrix is filled: for every neighborpair the position in th
 the value on this position of the GLCMatrix is increased +1
 */
 template <class T, size_t R>
-void GLCMFeatures2DFullMerge<T, R>::fill2DMatrices(boost::multi_array<T, R> inputMatrix, boost::multi_array<double, 2> &glcMatrix, int depth, int angle){
+void GLCMFeatures2DFullMerge<T, R>::fill2DMatrices(boost::multi_array<T, R> inputMatrix, boost::multi_array<float, 2> &glcMatrix, int depth, int angle){
     //vector in which all neihbor pairs are stored
     std::vector<std::pair<T, T> > neighbours;
     float weight;
@@ -156,11 +156,11 @@ std::vector<std::pair<T, T> > GLCMFeatures2DFullMerge<T, R>::getNeighbours2D(boo
 
 
 template <class T, size_t R>
-void GLCMFeatures2DFullMerge<T, R>::calculateAllGLCMFeatures2DFullMerge(GLCMFeatures2DFullMerge<T,R> &glcmFeatures, boost::multi_array<T, R> inputMatrix, float maxIntensity, vector<double> spacing, ConfigFile config){
+void GLCMFeatures2DFullMerge<T, R>::calculateAllGLCMFeatures2DFullMerge(GLCMFeatures2DFullMerge<T,R> &glcmFeatures, boost::multi_array<T, R> inputMatrix, float maxIntensity, vector<float> spacing, ConfigFile config){
     //get which norm should be used in the calculation of the GLCM features
     normGLCM = config.normGLCM;
     actualSpacing = spacing;
-    boost::multi_array<double,2> glcmFullMerge= glcmFeatures.calculateMatrix2DFullMerge(inputMatrix, maxIntensity);
+    boost::multi_array<float,2> glcmFullMerge= glcmFeatures.calculateMatrix2DFullMerge(inputMatrix, maxIntensity);
 	glcmFeatures.calculateJointMaximum(glcmFullMerge);
     glcmFeatures.calculateJointAverage(glcmFullMerge);
     glcmFeatures.calculateJointVariance(glcmFullMerge, this->jointAverage);
